services/search_service/requirements.txt
kafka-python
langchain
langchain-community
pyyaml
qdrant-client
httpx
fastapi
uvicorn
